Extracted inline styles from CliPage, ConnectionSelectorView, and ConnectionView components. Also added descriptive README.md file describing frontend flow
