+++ CLI/local-cli-fe-full/src/plugins/sshclient/styles/ConnectionView.css
@@ -1,0 +1,54 @@
+.connection-view {
+  flex: 1;
+  min-width: 0;
+  min-height: 0;
+  display: flex;
+  flex-direction: column;
+  padding: 16px;
+  box-sizing: border-box;
+  overflow: hidden;
+}
+
+.connection-view-scroll {
+  flex: 1;
+  min-height: 0;
+  overflow-y: auto;
+  overflow-x: hidden;
+  display: flex;
+  flex-direction: column;
+  gap: 12px;
+}
+
+.terminal-card {
+  display: flex;
+  flex-direction: column;
+  border: 1px solid #e2e8f0;
+  border-radius: 8px;
+  transition: background-color 0.2s;
+  overflow: hidden;
+}
+
+.terminal-card--single {
+  flex: 1 1 auto;
+}
+
+.terminal-card--multi {
+  flex: 0 0 auto;
+  height: calc(100vh - 200px);
+  min-height: calc(100vh - 200px);
+}
+
+.terminal-card--starred {
+  background-color: #fffbeb;
+}
+
+.terminal-card--unstarred {
+  background-color: white;
+}
+
+.terminal-card-body {
+  flex: 1;
+  min-height: 0;
+  display: flex;
+  flex-direction: column;
+}
