+++ CLI/local-cli-fe-full/src/plugins/sshclient/subcomponents/TerminalView.js
@@ -3,7 +3,7 @@ import { Terminal } from 'xterm';
 import { FitAddon } from 'xterm-addon-fit';
 import 'xterm/css/xterm.css';
 import { cliState } from '../state/state';
-import './TerminalView.css';
+import '../styles/TerminalView.css';
 
 const TerminalView = ({ id, host, user, credential, action, authType }) => {
   const terminalRef = useRef(null);
